+++ src/basic/capability-list.c
@@ -4,7 +4,7 @@
 
 #include "alloc-util.h"
 #include "bitfield.h"
-#include "cap-list.h"
+#include "capability-list.h"
 #include "capability-util.h"
 #include "extract-word.h"
 #include "log.h"
@@ -14,8 +14,8 @@
 
 static const struct capability_name* lookup_capability(register const char *str, register GPERF_LEN_TYPE len);
 
-#include "cap-from-name.inc"
-#include "cap-to-name.inc"
+#include "capability-from-name.inc"
+#include "capability-to-name.inc"
 
 const char* capability_to_name(int id) {
         if (id < 0)
