+++ src/basic/capability-to-name.awk
