+++ src/basic/generate-capability-list.sh
