+++ src/home/homectl.c
@@ -12,7 +12,7 @@
 #include "bus-error.h"
 #include "bus-locator.h"
 #include "bus-util.h"
-#include "cap-list.h"
+#include "capability-list.h"
 #include "capability-util.h"
 #include "cgroup-util.h"
 #include "creds-util.h"
