+++ src/libsystemd/sd-bus/bus-dump.c
@@ -8,7 +8,7 @@
 #include "bus-internal.h"
 #include "bus-message.h"
 #include "bus-type.h"
-#include "cap-list.h"
+#include "capability-list.h"
 #include "capability-util.h"
 #include "fileio.h"
 #include "format-util.h"
