+++ src/test/test-capability-util.c
